Added all knowledge to player area
When it becomes a player's turn they can see all the clues that they currently know not to be true. Clues the player has seen, either due to the card for that clue being in their hand or being shown that card by another person, are marked as red to show they are known to not be part of the game's solution.