Added ability to make an accusation at the end of a player's turn
The accusation is compared to the actual solution for the game. If the guess is correct then that player is the winner! If the guess is incorrect then their piece is removed from the game. Their future turns are skipped, but they will still present clues as normal during the suggestion phase if they hold the corresponding clue card.